Clarify what the ExternallyProvidedVertxService purpose is
